+++ packages/site-admin/src/stories/app.tsx
@@ -1,0 +1,14 @@
+/**
+ * Internal dependencies
+ */
+import { RouterProvider } from '../router';
+import { Layout } from './layout';
+import { routes } from './routes';
+
+export function App() {
+	return (
+		<RouterProvider routes={ routes } pathArg="page">
+			<Layout />
+		</RouterProvider>
+	);
+}