+++ packages/site-admin/src/stories/routes/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Internal dependencies
+ */
+import { componentsRoutes } from './components';
+import { homeRoutes } from './home';
+
+export const routes = [ ...homeRoutes, ...componentsRoutes ];