+++ packages/site-admin/src/stories/screens/index.ts
@@ -1,0 +1,3 @@
+export * from './home';
+export * from './docs-components';
+export * from './docs-routing-system';