+++ packages/site-admin/src/stories/sidebar/index.ts
@@ -1,0 +1,2 @@
+export * from './sidebar-components';
+export * from './sidebar-home';